Update Whoops\Run::handleException() for quitOnFinish option.
Update the control structure logic in Whoops\Run::handleException() to allow for the quitOnFinish option to be taken into account.
If the response of the Handler was not Handle::QUIT, then decide whether to quit depending on the value of $quitOnFinish (although this is ultimately decided by the value held in $allowQuit).