Composer: let version be determined by git tags
As per http://getcomposer.org/doc/02-libraries.md#specifying-the-version,
it is recommended to not put explicit version in the composer.json file.

Main advantage of that is that bumping a library version is as easy as adding
a tag, which is significantly different from editing a file and creating
a new commit with meaningless commit message.