TemplateHelper: fix escape with broken utf-8, close #163
If the source file that triggered an exception, had an illegal
UTF-8 sequence (because, for example, it wasn't UTF-8),
when PrettyPageHandler would try to show its source on the screen,
htmlspecialchars would return an empty string, resulting in no
source on the screen.

The workaround at least shows question marks instead of the original
characters. Guessing the correct encoding and reencoding that
to UTF-8 is too much bother at this point,
and the current approach still works well enough to see
where the error is coming from.