Changed key masking to display 'BLACKLISTED'
The reason for this is simply in the interest of security; by repeating '*' for each key's characters it makes it the slightest bit easier to tailor an attack by knowing each key's length. By displaying 'BLACKLISTED', you're both continuing to hide the specified superglobal key and mitigating a possible attack exploitation.