add previous exception messages to plain text output
Can be disabled via `PlainTextHandler::addPreviousToOutput(false)` and
adds lines between the exception and the trace containing the exception
and line of creator.

The method `Inspector::getFrames()` already appends all frames from the
previous exceptions. So there is no need in adding the trace for each
previous exception.

IMO the test tests something that it should not test: the output of the
line number. But that should be in a separate PR. I've at least added
constants for the line numbers so that you only have to change them once
for all tests.