Correct vscode open links
The format of links for vscode is currently set to:

    vscode://file/{$file}:{$line}

This leads to URIs such as:

    vscode://file//Users/example/git/example.php

However vscode expects URIs in the format:

    vscode://file/Users/example/git/example.php

That is to say that the URI should be immediately after the `file`.

The correct URI format is:

    vscode://file{$file}:{$line}

See https://github.com/microsoft/vscode/issues/197319 for more
information.